move slow tests to e2e
